Updated variable names to be consistent with Change servicenow plugin. Also added logic to let user if no service match, and if a match but no tickets
