+++ plugins/snow-incident-plugin/src/components/App.tsx
@@ -2,13 +2,13 @@ import type React from "react";
 import { PluginProvider } from "@cortexapps/plugin-core/components";
 import "../baseStyles.css";
 import ErrorBoundary from "./ErrorBoundary";
-import Snow from "./snow";
+import Incidents from "./Incidents";
 
 const App: React.FC = () => {
   return (
     <ErrorBoundary>
       <PluginProvider>
-        <Snow />
+        <Incidents />
       </PluginProvider>
     </ErrorBoundary>
   );
